+++ Server/src/main/java/com/example/Server/LogActionRepository.java
@@ -2,4 +2,4 @@ package com.example.Server;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LogActionRepository extends JpaRepository<LogActionEntity, Integer> {}
+public interface LogActionRepository extends JpaRepository<LogActionEntity, Long> {}
