+++ Server/src/main/java/com/example/Server/LogActionService.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface LogActionService {
     String createLogAction(LogActionEntity entity);
     List<LogActionEntity> readLogActions();
-    LogActionEntity readLogAction(Integer id);
-    String deleteLogAction(Integer id);
-    String updateLogAction(LogActionEntity entity, Integer id);
+    LogActionEntity readLogAction(Long id);
+    String deleteLogAction(Long id);
+    String updateLogAction(LogActionEntity entity, Long id);
 }
